reenabled notifications
