+++ src/controllers/mercadopagoController.js
@@ -1,5 +1,17 @@
 const mp = require("../utils/mercadoPago");
+require("dotenv").config();
+const nodemailer = require("nodemailer"); // Import Nodemailer
+const emailUser = process.env.EMAIL_USER;
+const emailPass = process.env.EMAIL_PASS;
 const { conn, Order, OrderDetail, Product } = require("../db");
+
+const transporter = nodemailer.createTransport({
+    service: "Gmail",
+    auth: {
+      user: emailUser,
+      pass: emailPass,
+    },
+});
 
 const processPayment = async (userId, info) => {
     try {
@@ -27,6 +39,7 @@ const processPayment = async (userId, info) => {
             transaction_amount: total
         };
 
+        const payerEmail = info.payer.email;
         const { response } = await mp.payment.save(preference)
 
         if (response.status == "approved" && response.status_detail == "accredited") {
@@ -54,6 +67,8 @@ const processPayment = async (userId, info) => {
             } catch(error) {
                 throw error
             }
+
+            sendPaymentNotification(payerEmail, total);
         }
 
         return { status: response.status, payment_id: response.id };
@@ -100,67 +115,21 @@ function validateStockFromCart(cart) {
     return invalid == null
 }
 
+const sendPaymentNotification = async (payerEmail, total) => {
+    const mailOptions = {
+      from: emailUser,
+      to: payerEmail,
+      subject: "Confirmacion de pago",
+      text: `Gracias por confiar en nosotros. El monto total de tu compra fue de: ${total} pesos`,
+    };
+  
+    try {
+      await transporter.sendMail(mailOptions);
+    } catch (error) {
+      console.error("Error sending payment notification email: " + error);
+    }
+  };
+
 module.exports = {
     processPayment
 }
-
-/*
-const mercadoPagoPayment = async (req, res) => {
-  try {
-    const {
-      description,
-      transaction_amount,
-      email,
-      payment_method_id,
-      token,
-      installments,
-    } = req.body;
-    const preference = {
-      items: [
-        {
-          title: description,
-          unit_price: transaction_amount,
-          quantity: 1,
-        },
-      ],
-      payer: {
-        email: email,
-      },
-    };
-    const response = await payment.preferences.create(preference);
-    console.log("respuesta de crear las preferencias", response);
-    const payment_data = {
-      token: token,
-      transaction_amount: transaction_amount,
-      payment_method_id: payment_method_id,
-      installments: installments,
-      payer: {
-        email: req.body.payer.email,
-      },
-    };
-    const emailDestinatario = payment_data.payer.email;
-    // Process the payment and send the email inside the .then block
-    payment.payment
-      .save(payment_data)
-      .then(async function (response) {
-        console.log("Respuesta de MercadoPago:", response.body);
-        console.log("Respuesta de MercadoPago:", req.body);
-        res.json({ status: response.status, payment_id: response.body.id });
-
-        // Send the payment notification email after a successful payment
-        await sendPaymentNotification(
-          emailDestinatario,
-          description,
-          transaction_amount
-        );
-      })
-      .catch(function (error) {
-        console.error("Error al procesar el pago:", error);
-        res.status(500).json({ message: "Error al procesar el pago" });
-      });
-  } catch (error) {
-    console.error("Error al procesar el pago:", error);
-    res.status(500).json({ message: "Error al procesar el pago" });
-  }
-};
-*/
